dialog tree extracted to separate file
